# kyros-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyros Backend",
    "description": "Wellness (Phase 1) + Clinic (Phase 2)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "summary": "Users List",
        "operationId": "users_list_admin_users_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "get": {
        "summary": "User Detail",
        "operationId": "user_detail_admin_users__user_id__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/edit": {
      "post": {
        "summary": "User Edit",
        "operationId": "user_edit_admin_users__user_id__edit_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_user_edit_admin_users__user_id__edit_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/deactivate": {
      "get": {
        "summary": "Deactivate Confirm Get",
        "operationId": "deactivate_confirm_get_admin_users__user_id__deactivate_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Deactivate Confirm Post",
        "operationId": "deactivate_confirm_post_admin_users__user_id__deactivate_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_deactivate_confirm_post_admin_users__user_id__deactivate_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/delete": {
      "get": {
        "summary": "Delete Confirm Get",
        "operationId": "delete_confirm_get_admin_users__user_id__delete_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Delete Confirm Post",
        "operationId": "delete_confirm_post_admin_users__user_id__delete_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_delete_confirm_post_admin_users__user_id__delete_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_deactivate_confirm_post_admin_users__user_id__deactivate_post": {
        "properties": {
          "confirmation": {
            "type": "string",
            "title": "Confirmation",
            "default": ""
          }
        },
        "type": "object",
        "title": "Body_deactivate_confirm_post_admin_users__user_id__deactivate_post"
      },
      "Body_delete_confirm_post_admin_users__user_id__delete_post": {
        "properties": {
          "confirmation": {
            "type": "string",
            "title": "Confirmation",
            "default": ""
          }
        },
        "type": "object",
        "title": "Body_delete_confirm_post_admin_users__user_id__delete_post"
      },
...
      "Body_user_edit_admin_users__user_id__edit_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "age": {
            "type": "string",
            "title": "Age",
            "default": ""
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "default": ""
          },
          "email": {
            "type": "string",
            "title": "Email",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": ""
          },
          "subscription_tier": {
            "type": "string",
            "title": "Subscription Tier",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          }
        },
        "type": "object",
        "title": "Body_user_edit_admin_users__user_id__edit_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}